전체 엔티티 틀
