+++ yyGang/src/main/java/com/beyond3/yyGang/answer/Answer.java
@@ -1,7 +1,8 @@
-package com.beyond3.yyGang;
+package com.beyond3.yyGang.answer;
 
 import com.beyond3.yyGang.q_board.QuestionBoard;
 import com.beyond3.yyGang.user.User;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -27,7 +28,9 @@ public class Answer {
     private String answerContents;
 
     @CreationTimestamp
+    @Column(updatable = false)
     private LocalDateTime answerDate;
+
     @UpdateTimestamp
     private LocalDateTime answerMdate;
