+++ yyGang/src/main/java/com/beyond3/yyGang/cart/Cart.java
@@ -1,6 +1,5 @@
 package com.beyond3.yyGang.cart;
 
-import com.beyond3.yyGang.CartOption;
 import com.beyond3.yyGang.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,12 +7,9 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToMany;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
-
-import java.util.List;
 
 @Entity
 @Getter
