+++ yyGang/src/main/java/com/beyond3/yyGang/order/Order.java
@@ -1,7 +1,5 @@
 package com.beyond3.yyGang.order;
 
-import com.beyond3.yyGang.OrderOption;
-import com.beyond3.yyGang.pay.Payment;
 import com.beyond3.yyGang.user.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,12 +8,8 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
-
-import java.util.List;
 
 @Entity
 @Getter
